fix: RM-1549 [iOS] crypto 提現時會造成crash

fix: RM-1549 #comment rootcause: api /each-crypto-limit回傳欄位格式為float與model 欄位格式int不符 \\solution: app model 更正為浮點數

Closes RM-1549

See merge request mobile/ktobet-asia-ios!878

+++ ktobet-asia-ios/domain/struct/ResponseData.swift
@@ -2000,8 +2000,8 @@ struct InternalMessageBean: Codable {
 struct CryptoLimitBean: Codable {
     let cryptoCurrency: Int
     let cryptoNetwork: Int
-    let max: Int
-    let min: Int
+    let max: Double
+    let min: Double
 }
 
 struct CryptoCurrencyBean: Codable {
